User Info and post update
Added user info page which allows the user to view posts they have added
and allows the user to remove those posts.

Added buttons to the for loop of the data recieved from the database
which allows the user to click on a button which returns info specific
to the given post. This will allow the user to see info about the
specific post.